prefetch-npm-deps: refactor
